Add support for continues and getting registers
Why: Both of these are important functionality for this library to have

How: Via the continue_forward and get_register functions. The function
signatures on both of these methods are subject to change as I think
both of them are lacking and do not properly capture the power of the
functions they are referring to.

Tags:

Footer
© 2022 GitHub, Inc.
Footer navigation