Refactored event loop & added proc map
This refactored the event loop to better reflect how rr originally
does it. I suspect that I had issues with continuing because of that.
This also added support for reading proc maps. It currently uses my
fork of procmaps but once my PR lands, we can switch to the official
one.