Added support for breakpoints and reverse continue
Why:
Self-explanatory

How:
Refactored the event loop in BinaryInterface.cc with run_req to properly
match how it originally is done. The way I have it set up may have
problems at the end of a program and for program re-starting but that
requires further testing.

Tags:

Footer
© 2022 GitHub, Inc.
Footer navigation